+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# 1.1.1
+
+- Ensured deterministic client generation
+- Added --info option to print information about the Prismarine package
+
 # 1.1.0
 
 - Integration with EasySAM